Restructure and add comments
