+++ src/components/template-import-export/operations/duplicateTemplate.ts
@@ -1,7 +1,6 @@
 import { ApiError } from '../../../ApiError'
 import db from '../databaseMethods'
-
-import { buildTemplateStructure } from '../buildTemplateStructure'
+import { buildTemplateStructure } from '../utilities'
 import { installTemplate } from './importTemplate'
 import { PgTemplate } from '../types'
