+++ src/components/template-import-export/operations/importTemplate.ts
@@ -3,7 +3,7 @@ import fsx from 'fs-extra'
 import semverCompare from 'semver/functions/compare'
 import { ApiError } from '../../../ApiError'
 import db from '../databaseMethods'
-import { filterModifiedData } from '../getDiff'
+import { filterModifiedData } from '../utilities'
 import { FILES_FOLDER, FILES_TEMP_FOLDER } from '../../../constants'
 import config from '../../../config'
 import {
@@ -19,7 +19,7 @@ import {
   PgTemplateCategory,
   TemplateStructure,
 } from '../types'
-import { hashFile, replaceForeignKeyRef } from '../updateHashes'
+import { hashFile, replaceForeignKeyRef } from '../utilities'
 
 interface InfoFile {
   timestamp: string
