+++ src/components/template-import-export/operations/index.ts
@@ -3,3 +3,4 @@ export * from './commitTemplate'
 export * from './duplicateTemplate'
 export * from './exportTemplate'
 export * from './importTemplate'
+export * from './linkingOperations'
