+++ src/components/template-import-export/utilities/buildTemplateStructure.ts
@@ -1,4 +1,13 @@
-import db from './databaseMethods'
+/**
+ * Pulls all template data, including "linked" items (e.g. data views, files,
+ * etc) from database and builds them into a single tree structure that is not
+ * dependent at all on database IDs.
+ *
+ * This is the structure that gets exported to file as JSON when exporting a
+ * template.
+ */
+
+import db from '../databaseMethods'
 import { getTemplateLinkedEntities } from './getTemplateLinkedEntities'
 import {
   CombinedLinkedEntities,
@@ -13,16 +22,7 @@ import {
   TemplateSection,
   TemplateStage,
   TemplateStructure,
-} from './types'
-
-/**
- * Pulls all template data, including "linked" items (e.g. data views, files,
- * etc) from database and builds them into a single tree structure that is not
- * dependent at all on database IDs.
- *
- * This is the structure that gets exported to file as JSON when exporting a
- * template.
- */
+} from '../types'
 
 export const buildTemplateStructure = async (template: PgTemplate) => {
   const { id: templateId, linked_entity_data, template_category_id, ...structure } = template
