+++ src/components/template-import-export/utilities/index.ts
@@ -1,0 +1,4 @@
+export * from './buildTemplateStructure'
+export * from './getDiff'
+export * from './getTemplateLinkedEntities'
+export * from './updateHashes'
